faq_assistant/requirements.txt
Flask
google-generativeai
python-dotenv
pymongo